Refactor product interfaces import and remove logs
Renamed interfaces.ts to interfaces/products.ts and updated imports in product-data.tsx accordingly. Removed unnecessary console.log statements from product data fetching functions for cleaner code.
